prime: first fully functional version of the app.
Initial version contains the base classes PrimeNumber
that takes in a number as string value and decides if it
is a prime number, returns a string result after processing.

Also a OutputChannel class to print the result either in the
console or to a file, decided based on a program argument
given by the user.

Basic test cases to cover the two classes.
App can be packaged to a jar and run via command line.